Move role related plans
