+++ src/query/legacy-planners/src/lib.rs
@@ -49,10 +49,6 @@ mod plan_partition;
 mod plan_projection;
 mod plan_read_datasource;
 mod plan_remote;
-mod plan_role_create;
-mod plan_role_drop;
-mod plan_role_grant;
-mod plan_role_revoke;
 mod plan_select;
 mod plan_setting;
 mod plan_sink;
@@ -159,10 +155,6 @@ pub use plan_read_datasource::SourceInfo;
 pub use plan_remote::RemotePlan;
 pub use plan_remote::V1RemotePlan;
 pub use plan_remote::V2RemotePlan;
-pub use plan_role_create::CreateRolePlan;
-pub use plan_role_drop::DropRolePlan;
-pub use plan_role_grant::GrantRolePlan;
-pub use plan_role_revoke::RevokeRolePlan;
 pub use plan_select::SelectPlan;
 pub use plan_setting::SettingPlan;
 pub use plan_setting::VarValue;
