+++ src/query/service/src/interpreters/interpreter_role_revoke.rs
@@ -15,8 +15,8 @@
 use std::sync::Arc;
 
 use common_exception::Result;
-use common_legacy_planners::RevokeRolePlan;
 use common_meta_types::PrincipalIdentity;
+use common_planner::plans::RevokeRolePlan;
 use common_users::UserApiProvider;
 
 use crate::interpreters::Interpreter;
