+++ src/query/service/src/sql/planner/binder/ddl/account.rs
@@ -21,14 +21,14 @@ use common_ast::ast::RevokeStmt;
 use common_exception::Result;
 use common_legacy_planners::AlterUserPlan;
 use common_legacy_planners::CreateUserPlan;
-use common_legacy_planners::GrantRolePlan;
-use common_legacy_planners::RevokeRolePlan;
 use common_meta_types::AuthInfo;
 use common_meta_types::GrantObject;
 use common_meta_types::UserOption;
 use common_meta_types::UserPrivilegeSet;
 use common_planner::plans::GrantPrivilegePlan;
+use common_planner::plans::GrantRolePlan;
 use common_planner::plans::RevokePrivilegePlan;
+use common_planner::plans::RevokeRolePlan;
 use common_users::UserApiProvider;
 
 use crate::sql::plans::Plan;
