+++ src/query/service/src/sql/planner/binder/mod.rs
@@ -26,15 +26,15 @@ use common_ast::UDFValidator;
 use common_datavalues::DataTypeImpl;
 use common_exception::Result;
 use common_legacy_planners::AlterUserUDFPlan;
-use common_legacy_planners::CreateRolePlan;
 use common_legacy_planners::CreateUserUDFPlan;
-use common_legacy_planners::DropRolePlan;
 use common_legacy_planners::DropUserPlan;
 use common_legacy_planners::DropUserStagePlan;
 use common_legacy_planners::DropUserUDFPlan;
 use common_legacy_planners::UseDatabasePlan;
 use common_meta_types::UserDefinedFunction;
 use common_planner::plans::CallPlan;
+use common_planner::plans::CreateRolePlan;
+use common_planner::plans::DropRolePlan;
 use common_planner::plans::ShowGrantsPlan;
 use common_planner::MetadataRef;
 pub use scalar::ScalarBinder;
